Move firebase config to src and update imports
